move tests into dir
