+++ src/Components/tests/Link.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import React from 'react'
 import { render } from 'react-testing-library'
-import Link from './Link'
+import Link from '../Link'
 
 describe('Link Component Test', () => {
   test('It renders', () => {
