+++ src/Components/tests/Positive.test.js
@@ -1,11 +1,11 @@
 /* eslint-disable */
 import React from 'react'
 import { render } from 'react-testing-library'
-import Header from './Header'
+import Positive from '../Positive'
 
 describe('Header Component Test', () => {
   test('It renders', () => {
-    const { getByText } = render(<Header />)
-    expect(getByText(/^Do I Need An Adapter?/)).toBeTruthy()
+    const { getByText } = render(<Positive />)
+    expect(getByText(/^You're All Set!/)).toBeTruthy()
   })
 })
